Refactored generation voting example code

+++ day2/section02-generation-voting/src/main/java/com/distribpatterns/generation/NextGenerationResponse.java
@@ -8,6 +8,6 @@ package com.distribpatterns.generation;
  * - Unique (quorum ensures no duplicates)
  * - Agreed upon by majority of replicas
  */
-public record NextGenerationResponse(int generation) {
+public record NextGenerationResponse(long generation) {
 }
