+++ day2/section02-generation-voting/src/main/java/com/distribpatterns/generation/PrepareRequest.java
@@ -10,6 +10,6 @@ package com.distribpatterns.generation;
  * 
  * Note: Uses strict > (not >=) to ensure uniqueness without server IDs.
  */
-public record PrepareRequest(int proposedGeneration) {
+public record PrepareRequest(long proposedGeneration) {
 }
